+++ src/ExceptionParser.php
@@ -44,6 +44,11 @@ final class ExceptionParser
     private ?Error $default = null;
 
     /**
+     * @var Closure|null
+     */
+    private ?Closure $accept = null;
+
+    /**
      * @var array
      */
     private array $pipes = [
@@ -61,8 +66,7 @@ final class ExceptionParser
      */
     public static function renderer(): Closure
     {
-        return static fn(Throwable $ex, $request) => self::make()
-            ->render($request, $ex);
+        return self::make()->renderable();
     }
 
     /**
@@ -170,6 +174,10 @@ final class ExceptionParser
      */
     public function isRenderable($request, Throwable $e): bool
     {
+        if (true === $this->mustAccept($e, $request)) {
+            return true;
+        }
+
         if ($e instanceof JsonApiException) {
             return true;
         }
@@ -197,6 +205,49 @@ final class ExceptionParser
             ->through($this->pipes)
             ->via('handle')
             ->then(fn() => new ErrorResponse($this->getDefaultError()));
+    }
+
+    /**
+     * @return $this
+     */
+    public function acceptsJson(): self
+    {
+        $this->accept = static fn($ex, $request) => $request->wantsJson();
+
+        return $this;
+    }
+
+    /**
+     * @param Closure $callback
+     * @return $this
+     */
+    public function accept(Closure $callback): self
+    {
+        $this->accept = $callback;
+
+        return $this;
+    }
+
+    /**
+     * @return Closure
+     */
+    public function renderable(): Closure
+    {
+        return fn(Throwable $ex, $request) => $this->render($request, $ex);
+    }
+
+    /**
+     * @param Throwable $ex
+     * @param $request
+     * @return bool
+     */
+    private function mustAccept(Throwable $ex, $request): bool
+    {
+        if ($this->accept) {
+            return ($this->accept)($ex, $request);
+        }
+
+        return false;
     }
 
     /**